Extend MAC collection in redfish
There is yet another flavor of MAC address organization.

So make sure we provide chances to hit it at least three of
the various ways redfish may choose to manifest mac addresses.

There are other paths to the data as well, but hopefully an
implementation at least implements one of the three.

Change-Id: Iaba947e2e16f14b6e7a94018beac046efd73bc25